src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/lifecycle-created-paste-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: document.registerElement is not a function
This test ensures that the lifecycle callbacks of editing-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL window.callbacksCalled should be X-FOO,DIV. Was .
FAIL window.callbacksCalled should be X-FOO,DIV,X-FOO,DIV. Was .
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
